requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.1
repath==0.9.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
Unipath==1.1